Fix Presentation Tests for height value
* Fix module for datasource

+++ pokemon-ui/src/main/java/com/pokemon/ui/model/PokemonInfoDetailUiModel.kt
@@ -3,6 +3,6 @@ package com.pokemon.ui.model
 data class PokemonInfoDetailUiModel (
     val name: String,
     val id: Int,
-    val height: Int,
+    val height: Double,
     val imageUrl: String
 )
